+++ lib/components/Input/HelperMessage.tsx
@@ -1,7 +1,7 @@
 import React, { ReactNode } from "react";
 import clsx from "clsx";
 import { InputVariants } from ".";
-import "./Input.scss";
+import "./HelperMessage.scss";
 
 export interface HelperMessageProps {
   error?: boolean;
@@ -9,11 +9,12 @@ export interface HelperMessageProps {
   variant?: InputVariants;
   disabled?: boolean;
 }
-
-const MessageVariant: Record<InputVariants, string> = {
-  general: "deriv-helper-message__general",
-  success: "deriv-helper-message__success",
-  error: "deriv-helper-message__error",
+type TMessageVariant = Exclude<InputVariants, "disabled">;
+const MessageVariant: Record<TMessageVariant, string> = {
+  general: "deriv-helper-message--general",
+  success: "deriv-helper-message--success",
+  warning: "deriv-helper-message--warning",
+  error: "deriv-helper-message--error",
 };
 
 const HelperMessage = ({
@@ -25,7 +26,8 @@ const HelperMessage = ({
   <p
     className={clsx("deriv-helper-message", {
       [MessageVariant["general"]]: disabled,
-      [MessageVariant[error ? "error" : variant]]: !disabled,
+      [MessageVariant[error ? "error" : (variant as TMessageVariant)]]:
+        !disabled,
     })}
   >
     {message}