+++ lib/components/Input/index.tsx
@@ -3,9 +3,13 @@ import clsx from "clsx";
 import HelperMessage from "./HelperMessage";
 import "./Input.scss";
 
-export type InputVariants = "general" | "success" | "error";
-  interface InputProps
-  extends Omit<ComponentProps<"input">, "style" | "placeholder"> {
+export type InputVariants =
+  | "general"
+  | "success"
+  | "error"
+  | "warning"
+  | "disabled";
+interface InputProps extends Omit<ComponentProps<"input">, "placeholder"> {
   label?: string;
   leftPlaceholder?: ReactNode;
   rightPlaceholder?: ReactNode;
@@ -15,15 +19,19 @@ export type InputVariants = "general" | "success" | "error";
 }
 
 const InputVariant: Record<InputVariants, string> = {
-  general: "deriv-input--field__general",
-  success: "deriv-input--field__success",
-  error: "deriv-input--field__error",
+  general: "deriv-input--general",
+  success: "deriv-input--success",
+  warning: "deriv-input--general",
+  error: "deriv-input--error",
+  disabled: "deriv-input--disabled",
 };
 
 const LabelVariant: Record<InputVariants, string> = {
-  general: "deriv-input--label__general",
-  success: "deriv-input--label__success",
-  error: "deriv-input--label__error",
+  general: "deriv-input__label--general",
+  success: "deriv-input__label--success",
+  warning: "deriv-input__label--general",
+  error: "deriv-input__label--error",
+  disabled: "deriv-input__label--disabled",
 };
 
 export const Input = ({
@@ -39,33 +47,42 @@ export const Input = ({
   ...rest
 }: InputProps) => {
   return (
-    <div className="deriv-input">
+    <div
+      className={clsx(
+        "deriv-input",
+        className,
+        InputVariant[error ? "error" : variant],
+        {
+          "deriv-input--disabled": disabled,
+        }
+      )}
+    >
       {leftPlaceholder && (
-        <div className="deriv-input--left-content">{leftPlaceholder}</div>
+        <div className="deriv-input__left-content">{leftPlaceholder}</div>
       )}
       <input
         placeholder={label}
-        className={clsx(
-          "deriv-input--field",
-          className,
-          InputVariant[error ? "error" : variant]
-        )}
+        className="deriv-input__field"
         id={id}
+        disabled={disabled}
         {...rest}
       />
       <label
         className={clsx(
-          "deriv-input--label",
-          LabelVariant[error ? "error" : variant]
+          "deriv-input__label",
+          LabelVariant[error ? "error" : variant],
+          {
+            "deriv-input--label--disabled": disabled,
+          }
         )}
         htmlFor={id}
       >
         {label}
       </label>
       {rightPlaceholder && (
-        <div className="deriv-input--right-content">{rightPlaceholder}</div>
+        <div className="deriv-input__right-content">{rightPlaceholder}</div>
       )}
-      <div className="deriv-input--helper-message">
+      <div className="deriv-input__helper-message">
         {message && (
           <HelperMessage
             message={message}