+++ lib/main.ts
@@ -1,8 +1,12 @@
 export { Button } from "./components/Button";
+export { Dropdown } from "./components/Dropdown";
 export { Input } from "./components/Input";
-export { PageLayout } from "./components/PageLayout";
 export { Loader } from "./components/Loader";
+export { PageLayout } from "./components/PageLayout";
 export { Tab, Tabs } from "./components/Tabs";
 export { Text } from "./components/Text";
-export { Tooltip } from "./components/Tooltip";
 export { ToggleSwitch } from "./components/ToggleSwitch";
+export { Tooltip } from "./components/Tooltip";
+export { useDevice } from "./hooks/useDevice";
+export { useOnClickOutside } from "./hooks/useOnClickOutside";
+export { PasswordInput } from "./components/PasswordInput";